ISSUE #3607 - move Route and AuthenticatedRoute to their own files
