+++ frontend/src/v5/services/routing/route.component.tsx
@@ -1,0 +1,78 @@
+/**
+ *  Copyright (C) 2022 3D Repo Ltd
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { useEffect } from 'react';
+import { useSelector } from 'react-redux';
+import { Route as BaseRoute } from 'react-router-dom';
+import { selectCurrentRevision } from '@/v4/modules/model/model.selectors';
+import { ProjectsHooksSelectors } from '../selectorsHooks/projectsSelectors.hooks';
+import { ContainersHooksSelectors } from '../selectorsHooks/containersSelectors.hooks';
+import { FederationsHooksSelectors } from '../selectorsHooks/federationsSelectors.hooks';
+import { formatMessage } from '../intl';
+
+export interface RouteProps {
+	computedMatch?: any;
+	title?: string;
+	path?: string;
+	exact?: boolean;
+	children?: any;
+}
+
+const DEFAULT_TITLE = formatMessage({ id: 'pageTitle.default', defaultMessage: '3D Repo | Online BIM collaboration platform' });
+const LOADING_TEXT = formatMessage({ id: 'pageTitle.loading', defaultMessage: 'Loading...' });
+
+const truncateValue = (str, max = 12) => {
+	if (!str) return null;
+	let truncatedStr = str.substring(0, max).trim();
+	if (str.length > truncatedStr.length) truncatedStr += '...';
+	return truncatedStr;
+};
+
+export const Route = ({ title, children, computedMatch: { params }, ...props }: RouteProps) => {
+	const projectName = ProjectsHooksSelectors.selectCurrentProjectName();
+	const containerName = ContainersHooksSelectors.selectContainerById(params.containerOrFederation)?.name;
+	const federationName = FederationsHooksSelectors.selectFederationById(params.containerOrFederation)?.name;
+	const containerOrFederationName = containerName || federationName || '';
+	const revisionTag = useSelector(selectCurrentRevision)?.tag || LOADING_TEXT;
+
+	const PARAM_REGEX = /:(\S+)/g;
+	const parseTitle = (string) => string.replace(PARAM_REGEX, (_, match) => {
+		switch (match) {
+			case 'project':
+				return truncateValue(projectName) || LOADING_TEXT;
+			case 'revision':
+				return params.revision ? `(${truncateValue(revisionTag)})` : '';
+			case 'containerOrFederation':
+				return truncateValue(containerOrFederationName) || LOADING_TEXT;
+			default:
+				return truncateValue(params[match]) || '';
+		}
+	});
+
+	const titleParsed = title ? parseTitle(title) : '';
+
+	useEffect(() => {
+		if (titleParsed) {
+			document.title = titleParsed;
+		}
+		return () => {
+			document.title = DEFAULT_TITLE;
+		};
+	}, [titleParsed]);
+
+	return <BaseRoute {...props}>{children}</BaseRoute>;
+};
