+++ frontend/src/v5/ui/routes/dashboard/index.tsx
@@ -18,11 +18,13 @@
 import { useRouteMatch, useLocation, Switch, Redirect } from 'react-router-dom';
 import { GlobalStyle } from '@/v5/ui/themes/global';
 import { formatMessage } from '@/v5/services/intl';
-import { AuthenticatedRoute, discardSlash, Route } from '@/v5/services/routing/routing';
+import { discardSlash } from '@/v5/services/routing/routing';
 import { NotFound } from '@/v5/ui/routes/notFound';
 import { DashboardLayout } from '@components/dashboard/dashboardLayout';
 import { ViewerCanvas } from '@/v4/routes/viewerCanvas';
 import { DashboardViewerLayout } from '@components/dashboard/dashboardViewerLayout/dashboardViewerLayout.component';
+import { Route } from '@/v5/services/routing/route.component';
+import { AuthenticatedRoute } from '@/v5/services/routing/authenticatedRoute.component';
 import { PasswordForgot } from '../login/passwordForgot';
 import { PasswordChange } from '../login/passwordChange';
 import { TeamspaceSelection } from '../teamspaceSelection';
