+++ frontend/src/v5/ui/routes/dashboard/projects/index.tsx
@@ -19,11 +19,12 @@ import { useEffect } from 'react';
 import { useParams } from 'react-router';
 import { useRouteMatch, Switch, Redirect } from 'react-router-dom';
 
-import { discardSlash, Route } from '@/v5/services/routing/routing';
+import { discardSlash } from '@/v5/services/routing/routing';
 import { DashboardParams, NOT_FOUND_ROUTE_PATH } from '@/v5/ui/routes/routes.constants';
 import { UsersActionsDispatchers } from '@/v5/services/actionsDispatchers/usersAction.dispatchers';
 import { DashboardFooter } from '@components/shared/dashboardFooter';
 import { formatMessage } from '@/v5/services/intl';
+import { Route } from '@/v5/services/routing/route.component';
 import { Federations } from './federations';
 import { Containers } from './containers';
 import { UserPermissions } from './userPermissions/userPermissions.component';
